Extract OtherPlayer and Inventory request handlers
