+++ src/js/index.js
@@ -65,6 +65,8 @@ import { handleGuildBattlegroundRequest } from './msg/GuildBattlegroundRequestHa
 import { guildExpeditionService } from './msg/GuildExpeditionService.js';
 import { handleGuildExpeditionServiceRequest } from './msg/GuildExpeditionRequestHandler.js';
 import { handleClanServiceRequest } from './msg/ClanServiceRequestHandler.js';
+import { handleInventoryServiceRequest } from './msg/InventoryRequestHandler.js';
+import { handleOtherPlayerServiceRequest } from './msg/OtherPlayerRequestHandler.js';
 import {
   otherPlayerService,
   otherPlayerServiceUpdateActions,
@@ -490,6 +492,23 @@ const addAvailablePacksFP = (amount) => {
   availablePacksFP += amount;
 };
 
+const setAvailablePacksFP = (amount) => {
+  availablePacksFP = amount;
+};
+
+const getPlayerID = () => {
+  return PlayerID;
+};
+
+const setPlayerState = (name, id) => {
+  PlayerName = name;
+  PlayerID = id;
+};
+
+const setCityProtections = (protections) => {
+  CityProtections = protections;
+};
+
 const getTotalAvailableFP = () => {
   return availablePacksFP + availableFP;
 };
@@ -835,105 +854,30 @@ function handleRequestFinished(request) {
               /*ConversationService */
               getConversation(msg);
             }
-          } else if (msg.requestClass == 'OtherPlayerService') {
-            if (msg.requestMethod == 'getOtherPlayerCityMapEntity') {
-              const selected = msg.responseData;
-              /*PlayerID*/
-              if (PlayerID != selected.player_id) PlayerName = '';
-              PlayerID = selected.player_id;
-              GBselected.id = selected.id;
-              GBselected.name = helper.fGBname(selected.cityentity_id);
-              GBselected.level = selected.level;
-              GBselected.max_level = selected.max_level;
-              GBselected.connected = selected.connected;
-              GBselected.total = selected.state.forge_points_for_level_up;
-              if (selected.state.invested_forge_points)
-                GBselected.current = selected.state.invested_forge_points;
-              else GBselected.current = 0;
-              console.debug(GBselected);
-            } else if (msg.requestMethod == 'getSocialList') {
-              /*PlayerID*/
-              otherPlayerServiceUpdateActions(msg.responseData);
-            } else if (msg.requestMethod == 'visitPlayer') {
-              if (showOptions.showVisit) {
-                clearVisitPlayer();
-                otherPlayerService(msg);
-              }
-            } else if (msg.requestMethod == 'rewardPlunder') {
-              //console.debug('cityentity_id:', msg.responseData.cityentity_id);
-              const rewards = msg.responseData[0].product.resources;
-              Object.keys(rewards).forEach((reward) => {
-                console.debug(reward);
-                var name = helper.fResourceShortName(reward);
-                var qty = rewards[reward];
-
-                if (!rewardsOtherPlayer[name]) rewardsOtherPlayer[name] = 0;
-                rewardsOtherPlayer[name] += qty;
-                console.debug(reward);
-              });
-
-              var reward = [];
-              reward.source = 'otherPlayer';
-              reward.name = '';
-              reward.amount = 0;
-
-              if (showOptions.showGErewards) {
-                showReward(reward);
-              }
-            } else if (msg.requestMethod == 'rewardResources') {
-              /*rewardPlunder */
-              const rewards = msg.responseData.resources;
-
-              Object.keys(rewards).forEach((reward) => {
-                // console.debug(reward);
-                var name = helper.fResourceShortName(reward);
-                var qty = rewards[reward];
-
-                if (!rewardsOtherPlayer[name]) rewardsOtherPlayer[name] = 0;
-                rewardsOtherPlayer[name] += qty;
-                // console.debug(reward);
-              });
-
-              var reward = [];
-              reward.source = 'otherPlayer';
-              reward.name = '';
-              reward.amount = 0;
-
-              if (showOptions.showGErewards) {
-                showReward(reward);
-              }
-
-              /*openChest */
-            } else if (msg.requestMethod == 'getCityProtections') {
-              /*City Protections*/
-              // console.debug('msg:', msg);
-              if (msg.responseData) CityProtections = msg.responseData;
-            }
-          } else if (msg.requestClass == 'InventoryService') {
-            if (msg.requestMethod == 'getGreatBuildings') {
-              /*InventoryService*/
-              // console.debug(msg.responseData);
-            } else if (msg.requestMethod == 'getItems') {
-              /*InventoryService*/
-              // 	console.debug("InventoryService",msg.responseData);
-              // console.debug(Object.keys(CityEntityDefs));
-              storage.set('CityEntityDefs', CityEntityDefs);
-              var forgePoints = 0;
-              if (msg.responseData.length) {
-                for (var j = 0; j < msg.responseData.length; j++) {
-                  if (msg.responseData[j].name == '10 Forge Points')
-                    forgePoints += msg.responseData[j].inStock * 10;
-                  else if (msg.responseData[j].name == '5 Forge Points')
-                    forgePoints += msg.responseData[j].inStock * 5;
-                  else if (msg.responseData[j].name == '2 Forge Points')
-                    forgePoints += msg.responseData[j].inStock * 2;
-                }
-                availablePacksFP = forgePoints;
-                if (document.getElementById('availableFPID'))
-                  document.getElementById('availableFPID').textContent =
-                    availablePacksFP + availableFP;
-              }
-            }
+          } else if (
+            handleOtherPlayerServiceRequest(msg, {
+              helper,
+              GBselected,
+              getPlayerID,
+              clearVisitPlayer,
+              otherPlayerService,
+              otherPlayerServiceUpdateActions,
+              showOptions,
+              rewardsOtherPlayer,
+              showReward,
+              setPlayerState,
+              setCityProtections,
+            })
+          ) {
+            // handled in module
+          } else if (
+            handleInventoryServiceRequest(msg, {
+              CityEntityDefs,
+              availableFP,
+              setAvailablePacksFP,
+            })
+          ) {
+            // handled in module
           } else if (
             msg.requestClass == 'ArmyUnitManagementService' &&
             msg.requestMethod == 'getArmyInfo'
